Ask openlayers to zoom back to original extent/zoom level as closely as it can when restoring snapshot

+++ Portal2/grails-app/domain/au/org/emii/portal/Snapshot.groovy
@@ -21,10 +21,10 @@ class Snapshot
 	String description
     
     //Bounding box
-    Integer minX
-    Integer minY
-    Integer maxX
-    Integer maxY
+    Float minX
+    Float minY
+    Float maxX
+    Float maxY
 	
 	static mapping =
 	{
